requirements.txt
# Python dependencies for 16 Fantasies 1 Cup
pandas>=1.5.0
requests>=2.28.0
jinja2>=3.1.0

# Data sources (optional)
kaggle>=1.5.12
PyYAML>=6.0.0

# Optional: For enhanced data processing
numpy>=1.21.0
scipy>=1.9.0

# Optional: For advanced analytics
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0